Remove the changeDim packet from respawn.
PE can't handle PC's dimension hack when respawning in the same world.
Since correct sky in the overworld is more important and only a
bukkit-relient hacky implementation can solve it currently, this is to
be removed.
The hacky bukkit implementation will probably be featured in a support
plugin if no more sense is blown into the mechanics.