Fix optionalchat -> string datawatcher converter.
String datwatcher object assumes empty strings as null, while
optionalchat has a separate boolean to indicate that. So if the
optionalchat string is not null, but the json -> legacy text converter
output an empty string, we add color code to that empty string, so
string datawatcher object won't assume it as null.