fix invisible armour stands not being invisible
WARNING! USES NMS! (couldn't find another way to do this)
it said that bedrock invisible entities cannot have custom names. We could try to replace this by checking if the entity is invisible and has a custom name, and if yes, use an area effect cloud (if you don't set a potion type, it's invisible), or horse variant -1, then set it to a custom name, so the hologram (that's probably what an invisible armor stand with a custom name is!) still works, and doesn't look weird.