Use own pool instead of netty recycler
Netty recycler can grow to insane sizes and that can't be controlled
from the code itself. Netty also supports recycling objects in a thread
that is different from allocating one, and growing the pool, but we
don't need this.
In our own pool we set the total max capacity to 200 (can be changed
using system property), which should be enough to hold ~200 players
(because usually transformation does 1 server packet -> 1 packet data).