Fix disconnect after handling handshake sometimes producing error
PS sends disconnect packet and closes connection after the send is
completed, which means that there can be a delay between those actions.
Client however sends handshake packet and instantly sends the loginstart
packet after that, but since the listener is still the handshake one,
handling of this packet would produce an error and cause an immediate
connection drop. To avoid that a login listener which doesn't actually
handle any login packets is now set first.